+++ src/components/beta/gux-rating/gux-rating-star/gux-rating-star.types.ts
@@ -1,0 +1,1 @@
+export type GuxRatingStarFill = 'empty' | 'half' | 'full';