+++ src/components/beta/gux-tag/tests/gux-tag.spec.ts
@@ -1,16 +1,23 @@
 import { newSpecPage } from '@stencil/core/testing';
 import { GuxTag } from '../gux-tag';
 
-describe('gux-tag-beta', () => {
-  it('should render', async () => {
-    const page = await newSpecPage({
-      components: [GuxTag],
-      html: `
-        <gux-tag-beta>Test</gux-tag-beta>
-      `,
-      language: 'en'
-    });
+const components = [GuxTag];
+const language = 'en';
 
-    expect(page.root).toMatchSnapshot();
+describe('gux-tag-beta', () => {
+  describe('#render', () => {
+    [
+      '<gux-tag-beta>default</gux-tag-beta>',
+      '<gux-tag-beta color="default">default (explicit)</gux-tag-beta>',
+      '<gux-tag-beta color="navy">navy</gux-tag-beta>',
+      '<gux-tag-beta icon="bolt" color="navy">navy</gux-tag-beta>',
+      '<gux-tag-beta icon="bolt" color="navy" value="3" removable>navy</gux-tag-beta>'
+    ].forEach((html, index) => {
+      it(`should render component as expected (${index + 1})`, async () => {
+        const page = await newSpecPage({ components, html, language });
+
+        expect(page.root).toMatchSnapshot();
+      });
+    });
   });
 });